Refactor Ajax class to use proper BogoDataWrapper instead of anonymous class
- Create BogoDataWrapper class to replace anonymous class
- Update Ajax get_bogo() method to return BogoDataWrapper instance
- Add BogoDataWrapper to BootstrapServiceProvider services
- Improve code structure and maintainability
- Remove anonymous class complexity
